+++ eras/shelley/test-suite/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Version history for `cardano-ledger-shelley-test`
 
-## 1.3.0.2
+## 1.4.0.0
 
-*
+* Remove `testBootstrapSpending`, `testBootstrapNotSpending`, `bootstrapHashTest` and `genSignature`
+* Moved `genBootstrapAddress` to `cardano-ledger-core:testlib:Test.Cardano.Ledger.Core.KeyPairs`
 
 ## 1.3.0.1